requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==2.1.1
dj-database-url==1.2.0
Django==4.1.6
django-admin-thumbnails==0.2.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
idna==3.4
Pillow==9.2.0
psycopg2-binary==2.9.5
pycparser==2.21
python-decouple==3.6
pytz==2023.3.post1
requests==2.28.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==1.26.16
whitenoise==6.2.0
